requirements.txt
requests
bs4
ebooklib
torch
tensorflow
transformers
python-dotenv
huggingface_hub